ci: give the two tightest files real headroom
The near-cap warning added alongside this immediately paid for itself: it named
`rankservice/mod.rs` at 495 and `metrics.rs` at 493 -- five and seven lines from
a cap that, when crossed, fails on whichever pull request happens to be open
next rather than the one responsible. That is exactly how main broke today.

`rankservice/mod.rs` is tight because of my own reservation-expiry work, so its
`Reservation` type and TTL move to `rankservice/reservation.rs`: 495 -> 471.
`metrics.rs` gives up its parser tests to `metrics/tests.rs`, the pattern this
tree uses everywhere: 493 -> 309. The sample fixture stays where it is and the
include path walks up one level, which is the only thing the move had to get
right.

No behaviour change: 11 suites green, clippy and fmt clean, and the tests that
moved still run (24 in the metrics module, 149 in atlasctl).

Not fixed here, and now visible in every run instead of only in a failure:
`docker/translate/tests.rs` sits at 494 with six lines to spare. It is the next
one, and whoever adds to it will be told before they land.
