+++ crates/atlasctl/src/rankservice/mod.rs
@@ -36,33 +36,6 @@ use atlasctl_core::settings;
 use std::collections::BTreeMap;
 use std::sync::{Arc, Mutex};
 
-/// A rendered rank, held between prepare and commit.
-struct Reservation {
-    epoch: String,
-    recipe: String,
-    plan: atlasctl_core::docker::LaunchPlan,
-    /// When this hold stops binding, computed FORWARD from creation.
-    ///
-    /// Without an expiry a reservation is immortal: the head that made it can
-    /// close its tab, crash, or restart for an upgrade, and nothing here ever
-    /// releases it -- so every later cluster launch on this machine is refused
-    /// until someone restarts the agent by hand. On a fleet, all of them.
-    ///
-    /// Stored as a deadline rather than a creation time so that nothing ever
-    /// computes `Instant::now() - d`, which PANICS on Windows when `d` exceeds
-    /// the time since boot. A CI runner is minutes old; that panic broke main.
-    expires: std::time::Instant,
-}
-
-/// How long a reservation outlives the head that made it.
-///
-/// It only has to cover prepare -> commit, which is a few round trips plus
-/// whatever `docker rm -f` takes; commit consumes the reservation, so a live
-/// launch is never at risk from this. Ten minutes is far past that and still
-/// short enough that an operator who lost a head can simply try again rather
-/// than ssh to every machine.
-const RESERVATION_TTL: std::time::Duration = std::time::Duration::from_secs(600);
-
 /// What this machine can do, as opposed to what it ships.
 ///
 /// Grouped rather than passed loose: these three are one question — whether
@@ -488,6 +461,9 @@ pub(crate) fn is_ours(container: &str) -> bool {
             .chars()
             .all(|c| c.is_ascii_lowercase() || c.is_ascii_digit() || c == '.' || c == '-')
 }
+
+mod reservation;
+pub(crate) use reservation::*;
 
 #[cfg(test)]
 mod commit_tests;
